Code refactor (#169)

+++ src/main/kotlin/server/endpoints/create/NewSectorEndpoint.kt
@@ -8,11 +8,11 @@ import database.entity.Zone
 import database.entity.info.LastUpdate
 import distribution.Notifier
 import io.ktor.http.HttpStatusCode
-import io.ktor.server.application.ApplicationCall
-import io.ktor.util.pipeline.PipelineContext
+import io.ktor.server.routing.RoutingContext
 import java.io.File
 import kotlinx.serialization.json.Json
 import server.endpoints.SecureEndpointBase
+import server.error.Error
 import server.error.Errors.MissingData
 import server.error.Errors.ParentNotFound
 import server.request.save
@@ -23,7 +23,7 @@ import storage.Storage
 import utils.isAnyNull
 
 object NewSectorEndpoint : SecureEndpointBase("/sector") {
-    override suspend fun PipelineContext<Unit, ApplicationCall>.endpoint() {
+    override suspend fun RoutingContext.endpoint() {
         var displayName: String? = null
         var point: LatLng? = null
         var kidsApt: Boolean? = null
@@ -35,6 +35,7 @@ object NewSectorEndpoint : SecureEndpointBase("/sector") {
         var imageFile: File? = null
         var gpxFile: File? = null
 
+        var error: Error? = null
         receiveMultipart(
             forEachFormItem = { partData ->
                 when (partData.name) {
@@ -45,7 +46,7 @@ object NewSectorEndpoint : SecureEndpointBase("/sector") {
                     "walkingTime" -> walkingTime = partData.value.toUIntOrNull()
                     "weight" -> weight = partData.value
                     "zone" -> ServerDatabase.instance.query {
-                        zone = Zone.findById(partData.value.toInt()) ?: return@query respondFailure(ParentNotFound)
+                        zone = Zone.findById(partData.value.toInt()) ?: return@query ParentNotFound.let { error = it }
                     }
                 }
             },
@@ -56,6 +57,10 @@ object NewSectorEndpoint : SecureEndpointBase("/sector") {
                 }
             }
         )
+        if (error != null) {
+            call.respondFailure(error)
+            return
+        }
 
         if (isAnyNull(displayName, imageFile, kidsApt, sunTime, zone)) {
             imageFile?.delete()
